Make socket mark functions linux exclusive
